Extract witness functions to separate file
