+++ Source/Concurrency/CivlUtil.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.Boogie;
 
 namespace Microsoft.Boogie
@@ -29,6 +30,20 @@ namespace Microsoft.Boogie
         public static OldExpr Old(Expr expr)
         {
             return new OldExpr(Token.NoToken, expr);
+        }
+    }
+
+    public static class LinqExtensions
+    {
+        public static IEnumerable<IEnumerable<T>> CartesianProduct<T>(this IEnumerable<IEnumerable<T>> sequences)
+        {
+            IEnumerable<IEnumerable<T>> emptyProduct = new[] { Enumerable.Empty<T>() };
+            return sequences.Aggregate(
+                emptyProduct,
+                (accumulator, sequence) =>
+                from acc in accumulator
+                from item in sequence
+                select acc.Concat(new[] { item }));
         }
     }
 }
