+++ Source/Concurrency/Concurrency.csproj
@@ -85,6 +85,7 @@
     <Compile Include="YieldTypeChecker.cs" />
     <Compile Include="CivlUtil.cs" />
     <Compile Include="NewTransitionRelationComputation.cs" />
+    <Compile Include="Witnesses.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
